% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.r
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data}
\usage{
aggregate_data(Data, formula, FUN1 = median, FUN2 = function(x) {    
  sd(x)/sqrt(length(x)) })
}
\arguments{
\item{Data}{An output from \code{read_growth_plates}, or multiple outputs
concatenated with \code{rbind}.}

\item{formula}{A formula interface indicating how to aggregate the data.
See help for \code{aggregate}, for more details.}

\item{FUN1}{a function perform on aggregated data}

\item{FUN2}{a second function to perform on aggregated data}
}
\value{
Aggregated data.frame. For FUN2 the results will be in column
OD600.se
}
\description{
Simply perfromfs aggregate twice with a different function each time
}
\details{
Takes output from \code{read_growth_plates} and aggregates readings
from multiple replicates. Can handle an arbitrary number of replicates

This function is a wrapped for \code{aggregate}, see help of that
function for more details
}
\author{
Sur Herrera Paredes
}

