% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profiles.r
\name{cluster_profiles}
\alias{cluster_profiles}
\title{Cluster growth curves}
\usage{
cluster_profiles(Data, formula, value.var = "OD600", method = "complete")
}
\arguments{
\item{Data}{A data.frame}

\item{formula}{A formula for \link{acast} that will turn
the data.frame in Data into a matrix. This formula should define
the individual growth curves}

\item{value.var}{column namme containing growth curve values.
Value to be cast by \link{acast}.}

\item{method}{Method for hierarchical clustering to be
used by \link{hclust}.}
}
\value{
An hclus object
}
\description{
Performs hierarchical clusrering of growth curves based
on correlation distance. Cluster growth profiles of
different strains.
}
\details{
This funciton uses acast to convert an object from aggregate_data
into a matrix. Then uses pearson correlation to calculate a
dissimilarity betwenn profiles, where d_{ij} = 1 - p_{ij},
where d_{ij} is the distance between profiles i and j,
and p_{ij} is the pearson correlation between samples i and j.
Finaly ot performs hierarchical clustering on the dissimilarity matrix.
}
\author{
Sur Herrera Paredes
}
\seealso{
\link{hclust} \link{acast}
}

