% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgca.R
\name{pgca}
\alias{pgca}
\alias{pgca.data}
\alias{pgca.files}
\title{Link Protein Groups Created from MS/MS Data}
\usage{
pgca(dir, col.mapping)

pgca.data(..., col.mapping)

pgca.files(..., col.mapping)
}
\arguments{
\item{dir}{path to the directory.}

\item{col.mapping}{column mapping (see Details).}

\item{...}{arbitrary number of \code{data.frame}s or filenames.}
}
\value{
An object of type \code{accession.dict}.
}
\description{
Build a dictionary for protein groups from MS/MS data.
Details of the algorithm can be found in XX, YY (2017).
}
\section{Functions}{
\itemize{
\item \code{pgca.data}: Use data given as \code{data.frame}s.

\item \code{pgca.files}: Read in data from the given files
}}

\examples{
# Build the dictionary from all files in a directory
# and specifying the column "Gene_Symbol" holds the `gene.symbol`.
dict.dir <- pgca(
         system.file("extdata", package = "pgca"),
         col.mapping = c(gene.symbol = "Gene_Symbol")
)

# Build the dictionary from a list of files
dict.files <- pgca.files(
     system.file("extdata", "accs_no_1947.txt", package = "pgca"),
     system.file("extdata", "accs_no_2007.txt", package = "pgca"),
     system.file("extdata", "accs_no_2047.txt", package = "pgca"),
     col.mapping = c(gene.symbol = "Gene_Symbol")
)

# Build the dictionary from already read-in data.frames
df.1947 <- read.delim(system.file("extdata", "accs_no_1947.txt",
                                  package = "pgca"))
df.2007 <- read.delim(system.file("extdata", "accs_no_2007.txt",
                                  package = "pgca"))
dict.data <- pgca.data(df.1947, df.2007,
                       col.mapping = c(gene.symbol = "Gene_Symbol"))

}
\references{
XX, YY (2017). xxx.
}
\seealso{
\code{\link{translate}} to apply the dictionary to the data files
     and \code{\link{save.dict}} to save the dictionary itself.
}
