% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_profiles.r
\name{plot_cluster_profiles}
\alias{plot_cluster_profiles}
\title{Plot cluster profiles}
\usage{
plot_cluster_profiles(clustering, Data, h)
}
\arguments{
\item{clustering}{An hclus object.}

\item{Data}{A data.frame with the columns define by \code{\link{read_growth_plates}}.}

\item{h}{Height to cut the clustering dedrogram. To be used by \code{\link{cutree}}.}
}
\value{
A ggplot2 object from \code{\link{plot_growth_curves}}
}
\description{
Aggregate wells from the same cluster and plot them.
}
\author{
Sur Herrera Paredes
}

