% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strain_cluster_profile.r
\name{plot_strain_cluster_profile}
\alias{plot_strain_cluster_profile}
\title{Plot strain cluster profile}
\usage{
plot_strain_cluster_profile(clustering, Data, h, strain)
}
\arguments{
\item{clustering}{Result from \link{cluster_profiles}}

\item{Data}{A data.frame.}

\item{h}{Clustering height threshold for making groups.}

\item{strain}{focal strain ID.}
}
\value{
A ggplot2 plot
}
\description{
Plots growth curves of cluster containing a given strain
at a given height
}
\author{
Sur Herrera Paredes
}

