% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save.dict}
\alias{save.dict}
\title{Write PGCA dictionary to a text file}
\usage{
save.dict(dict, file = stop("`file` must be specified"), ...)
}
\arguments{
\item{dict}{a PGCA dictionary.}

\item{file}{either a character string naming a file or a
\code{\link[base]{connection}} open for writing.
\code{""} indicates output to the console.}

\item{...}{further arguments passed to \code{\link[utils]{write.table}}.}
}
\value{
This function returns \code{NULL} invisibly.
}
\description{
Write the dictionary to a text file using the
\code{\link[utils]{write.table}} function. By default, a tab-separated file
is written, but this can be changed
by changing the arguments to \code{\link[utils]{write.table}}.
}
\examples{
# Build accession dictionary from all files in a directory
dict <- pgca(
         system.file("extdata", package = "pgca"),
         col.mapping = c(gene.symbol = "Gene_Symbol")
)

\dontrun{
# Save dictionary to a file
save.dict(dict, file = "dictionary.txt")

# Change the separator string to a tab
save.dict(dict, file = "dictionary.txt", sep = "\\t")
}
}
\seealso{
\code{\link{pgca}} to create the dictionary, and
     \code{\link{translate}} to saved translated data files.
}
