% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{translate}
\alias{translate}
\title{Apply a PGCA dictionary to data files}
\usage{
translate(dict, input, out.dir, out.suffix = "", out.prefix = "",
  col.mapping, out.pg.col = "PGC")
}
\arguments{
\item{dict}{the PGCA dictionary to use.}

\item{input}{input directory, files, or data (see details).}

\item{out.dir}{the directory to save the translated files in (see details).
If \code{NULL}, the translated data frames will be returned directly.}

\item{out.suffix, out.prefix}{suffix and prefix that will be added to the
translated files.}

\item{col.mapping}{the column mapping for the input files. Defaults to the
same as used to build the dictionary.}

\item{out.pg.col}{the name of the column to store the protein group.}
}
\value{
Either a list of \code{data.frame}s or nothing (see details).
}
\description{
Apply the dictionary to the data files and write the translated files to
disk.
}
\details{
The dictionary is applied to the data specified in the \code{input} argument.
If the \code{input} argument is missing, the dictionary is applied to the
files used to create the dictionary. The \code{input} argument can either be
a directory, a character vector of file names, or a list of
\code{data.frames}s.

If the output directory \code{out.dir} is specified, the translated files
will be saved in this directory. Otherwise the files will be written to the
same directory as the input files. The function will not overwrite existing
files and will fail if the files already exist. Parameters \code{out.suffix}
and \code{out.prefix} can be used to ensure unique new file names.
In case the input is a list of \code{data.frame}s and no output
directory is specified, the \code{data.frame}s will be translated and
returned as a list. The function will also return the translated
\code{data.frame}s as list if the \code{out.dir = NULL}.
}
\examples{
# Build PGCA dictionary from all files in a directory
dict <- pgca(
         system.file("extdata", package = "pgca"),
         col.mapping = c(gene.symbol = "Gene_Symbol")
)

# Translate all files in the directory and return as a list of data.frames
trans <- translate(dict, input = system.file("extdata", package = "pgca"),
                   out.dir = NULL)

# Translate only some files in the directory and return as a list of
# data.frames
trans <- translate(dict, input = c(
                       system.file("extdata", "accs_no_1947.txt",
                                   package = "pgca"),
                       system.file("extdata", "accs_no_2007.txt",
                                   package = "pgca")
                   ), out.dir = NULL)
str(trans)

\dontrun{
# Translate all files in the directory and save to another directory
translate(dict, input = system.file("extdata", package = "pgca"),
          out.dir = "translated")

# Translate all files in the directory, save to the same directory, but
# add a prefix to the files
translate(dict, input = system.file("extdata", package = "pgca"),
          out.prefix = "translated_")
}

}
\seealso{
\code{\link{pgca}} to create the dictionary
}
